realtime_analysis/requirements.txt
requests>=2.31.0
gtfs-realtime-bindings>=1.0.0
psycopg2-binary>=2.9.9
pandas>=2.1.0
python-dotenv>=1.0.0
plotly>=5.17.0
kaleido>=0.2.1
sqlalchemy>=2.0.0


